+++ commands/help/reload.js
@@ -30,7 +30,7 @@ module.exports = {
 
         const commandFolders = fs.readdirSync(commands_path);
         const folderName = commandFolders.find(folder => fs.readdirSync(`${commands_path}/${folder}`).includes(`${commandName}.js`));
-        const name = msg.client.mod_getter.get_name(command)
+        const name = msg.client.mods.name.get(null, command)
 
         delete require.cache[require.resolve(`../${folderName}/${name}.js`)];
