+++ js/slash_commands/slasher.js
@@ -41,8 +41,8 @@ async function get_slash_commands(client) {
 
 async function create_slash_command(client, command) {
     const fake_msg = {author: {id: "-1", username: "slash_command"}, client: client}
-    const name = client.mod_getter.get_name(command)
-    const description = await client.mod_getter.get_description(fake_msg, command)
+    const name = client.mods.name.get(null, command)
+    const description = await client.mods.description.get(fake_msg, command)
 
     const data = new SlashCommandBuilder()
         .setName(name)
@@ -60,9 +60,9 @@ async function create_slash_command(client, command) {
 
 async function create_options(client, command) {
     const options = []
-    const name = client.mod_getter.get_name(command)
-    const args_min_length = client.mod_getter.get_args_min_length(command)
-    const args_max_length = client.mod_getter.get_args_max_length(command)
+    const name = client.mods.name.get(null, command)
+    const args_min_length = client.mods.args_min_length.get(null, command)
+    const args_max_length = client.mods.args_max_length.get(null, command)
 
     if (fs.existsSync(`./js/slash_commands/option_models/${name}.js`)) {
         const option_model = require(`./option_models/${name}.js`)
@@ -107,7 +107,7 @@ function create_option(name, description, required, choices) {
 // Checker
 // ----------------------------
 function check_slash(client, command) {
-    return client.mod_getter.get_enable_slash(command)
+    return client.mods.enable_slash.get(null, command)
 }
 // ----------------------------
